A programmer's client (a local cinema) website got hacked, with racist messages posted. The policy suspected the programmer hacked the site, and got a warrant to seize the programmer's computer. Then they sent **21 armed polices**, 3 of them to seize the computer, and **18 of them as witness**.
On early morning, these polices **broke in** the programmer's home, and **shot dead** the programmer when he is talking to his lawyer on the phone (**4 shots on chest and face**).
And the **18 witness polices** witnessed that the programmer fired a crossbow bolt at one of them (that police was protected by Kevlar thus slightly injured).

This is *not* something happening in a magic realism novel or an absurdist  movie. This is what happened on **December 23rd 2016**, **London, Ontario, Canada**.

The programmer is **Sam Maloney**, the creator of [morphis](https://morph.is/) (a free open source p2p distributed datastore), a husband and a father of two children.

News reports:

1. [first news report on nationalpost.com](http://news.nationalpost.com/news/canada/lawyer-tells-client-on-phone-during-predawn-raid-theyre-going-to-shoot-you-sam-put-your-hands-up)
2. [more recent report on Motherboard](https://motherboard.vice.com/read/why-did-police-kill-an-alleged-small-time-hacker-canada-sam-maloney) ([ZeroNet mirror](http://127.0.0.1:43110/1uEc35aRpkDgVmJ35jcMEHm4D2JCcEejp/motherboard/))