One of my favorite poem of Li Ch'ing-chao (李清照) is "As in a Dream" (如夢令）.
Light rain and gusty wind last night. （昨夜雨疏風驟）
Sound sleep did not dispel the slight drunkenness. （濃睡不消殘酒）
I ask the one rolling up the curtain,　（試問卷簾人）
but she answers: "The cherry-apple trees are the same."　（卻道海棠依舊）
"Don't you know? Don't you know? The red should languish while the green should plump."　（知否，知否，應是綠肥紅瘦）

(I am a native speaker of Chinese but not a native speaker of English. I tried my best to translate it. Any feedback on translation is welcome.)

"The red should languish while the green should plump" reveals the tenderness of the poet, while "should" implies the firmness of the poet.

"Don't you know? Don't you know?" That is the reality distortion field of a groggy poet.