My personal choices to score movies:
1. WTF
2. I do not want to waste my time on this
3. to kill some time
4. I probably will not watch it again, but I am not object to
5. I will watch it again

So forgive me for being harsh to a lot of movies,
I am poor, thus:

- I am busy with earning a living, not having time to watch so many movies;
- I have to limit the money spending on extra movie tickets.

However, I still watch quite a lot of movies after all.
Thanks to my just the right amount of poverty, I can afford buying movie tickets, and cannot afford some other more expensive forms of entertainment. In fact I am not sure, movie is entertainment to me. Maybe it is rather a form of escaping of the desert of reality? Or, maybe all forms of entertainment are ways of escaping to me?