Just came across a video titled [Learn Python Programming Language in 2 Hours][video].
The speaker argues that the first reason to use Python is:

> Solve complex problems in **less time** and **fewer lines of code** 

[video]: https://youtu.be/yE9v9rt6ziw

with the following example:

```
C# str.Substring(0, 3)
JavaScript str.substr(0, 3)
Python str[0:3]
```

I do not want to talk about technical details like C# 8.0 supports slicing notation (`str[0..3]`) and `substr` is a legacy function in JavaScript (use `slice` instead). What I want to say is that all of the three languages take one line of code to solve the problem, and with a decent editor/IDE, all three lines take equal time to type. Even assuming we all do programming with a basic text editor like Notepad on Windows or Leafpad in Linux, this example still has nothing to do with solving complex problems in less time. This problem is not complex at all. And most time of programming is not spent on typing.

Currently this video has 411k views, 8.1k upvotes (less than 200 downvotes) on YouTube. I think this is a good example of why learning to code is hard: so many poor and popular teaching materials on Internet.

![attached image](https://mmap.page/log/1555771818.jpg)