Just tried the first three quizs of [Python introduction course on stepik][stepik], which is included in JetBrains' PyCharm Educational (I tried it with the EduTools plugin on IntelliJ).
[stepik]: https://stepik.org/course/238

The first quiz is:

```python
print("Hello, world! My name is type your name")
```

So easy.

```python
print("Hello, world! My name is weakish.")
```

Failed! It turns out that the auto answer detection system thinks it is wired that my name is `weakish.`! Obviously the first quiz is to teach you just follow the instruction brainlessly. "My name ..." starts with a capital character, so it is a sentence. It is reasonable to end a sentence with a period (`.`), although this will fail the quiz.

The second quiz introduces comments, without telling the learner comment is used for what. 

And the third quiz introduces variable, but its first line is:

```python
a = b = 2
```

What? Introducing chained assignment before normal variable assignment (`a = 2`)? 

Chained assignment is not often used in practice. And I do not think it is necessary to introducing it to beginners.

In fact it is not only unnecessary, but confusing.

A beginner may guess that `a = b = 2` is equivalent to:

```python
a = 2
b = 2
```

For a value `2`, they are indeed equivalent, but:

```python
import random
a = b = random.random()
```

is not equivalent to:

```python
a = random.random()
b = random.random()
```

And programmers from other languages such as C and JavaScript, may think `a = b = 2` is equivalent to `a = (b = 2)` where `b = 2` evaluates to `2`. But this is not true in Python.

And the course mentioned nothing about these quirks.

I guess the author may think `a = 2` is too boringly obvious to introduce, thus turns to introducing chain assignment directly. No! In fact, considering that in Python variables are reassignable, they are not boring for beginners at all.

Considering the following lines:

```python
a = 2
a = 3
```

This is not boring for a beginner. How can `a` be equal to both `2` and `3`? Seems mathematically impossible!

This is what the author really have to talk about, not chain assignment.

I myself learned Python via reading the A Byte of Python book years ago. I am not sure if I should recommend it to beginners. Compared to [HtDP] for the exotic Scheme language and the outdated [Why's Guide to Ruby][why], it is somehow a boring book. And Python's official tutorial is well written. But at least it does not have such flaws.

[HtDP]: https://htdp.org/
[why]: https://poignant.guide/