The three most popular style guides in JavaScript community (airbnb, standard, google) are all unnecessarily lengthy.
For example, they all contain a lot of rules on spacing. Is spacing that important? If so, why not just use prettier? If not, why bother to list them in a style guide?

Another example is they all tell you to write `let arr = [];` instead of `let arr = new Array()`. I personally tend to write `[]` but if a ~~Java~~ developer writes `new Array()` instead, what's wrong with that? It is totally harmless.

Style guides should keep focus on a dozen of matters truly important, in other words, pitfalls of JavaScript design, e.g. using `===` instead of `==`.