Q: Why I do not have faith in God?
A: As Larry Wall paraphrased, faith is just two bits away:

> 1. God exists, and
> 2. God is good to people who really look for him

I am not able to prove the first bit is true but let me just assume the first bit is true.

The second bit is I have to really look for the God. Jesus said you only have to have faith the size of a mustard seed. But no matter how small the size of the faith is, it is **still an opt-in option**. If God is really good, then why faith is **not an opt-out option**?

Because we are sinners. And the sensible default option for sinners is no faith and going to hell. Salvation is a **gift** from the God.

However, **a gift is something declinable by its very definition**. Thus I can choose to **decline this gift and keep on the default option**. I did so.

Although I do not have faith in God, I am still grateful to God. **Thank God for allowing such a sinner as I to have the free will to decline his gift**.

And what if the first bit is not true? Then I would like to thank god for being nonexistent.